Configurable and flexible biome lookups, allows for modding capability
 + Switch from fixed match-statement biome lookup to flexible grid approach
   * Removes dependency on nightly (yay)
   * Also removes the deprecated use of literals in match statements
   * Better modding capability
 * General code cleanup to reduce amount of warnings
 + Add Segmented for the flexible grid implementation, replaces commented prototype for Grid
 - Remove set_block/get_block from anvil::Section as they are unimplemnted
 + Add All/None matchers for BlockMatcher
 * Turn off dead_code warnings for now, as there are 1800 lines of warnings currently

+++ src/noise_field/height.rs
@@ -77,6 +77,6 @@ impl HeightSource {
 /// 3 => 10
 /// 4 => 13
 /// ```
-pub fn lerp_to_layer(lerp: Vector2<i32>) -> Vector2<i32> {
+pub fn lerp_to_layer(lerp: Vector2<usize>) -> Vector2<usize> {
 	lerp.map(|x| x*3 + 1)
 }
